+++ examples/calculator-product/apps/calculator_api/test/calculator_gherkin_http_test.dart
@@ -2,9 +2,7 @@ import 'dart:io';
 
 import 'package:calculator_api/calculator_api.dart';
 import 'package:calculator_contracts/calculator_contracts.dart';
-import 'package:zuke_frontend/zuke_frontend.dart';
-import 'package:zuke_runner/zuke_runner.dart';
-import 'package:zuke_runner/http.dart';
+import 'package:zuke/http.dart';
 import 'package:test/test.dart';
 
 void main() {